fixed workgraph and submission
